moved base smuggle function to core module
